requirements.txt
datasets==2.19.0
einops==0.4.0
gpustat==1.1
hydra-core==1.1.1
higher==0.2.1
importlib-metadata==6.3.0
matplotlib==3.5.1
nltk==3.6.5
numpy==1.22.1
omegaconf==2.1.1
pandas==1.4.0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
tokenizers
torch==2.0.1
tqdm==4.62.3
transformers==4.40.0
openai==0.27.9
peft==0.7.1
timm==0.9.7
iopath==0.1.10
opencv-python==4.8.0.76
fairscale==0.4.13
